restygwt 271: don't find subtypes of concrete leaf types
a leaf type won't have any subtypes, so just return the type itself.
this fixes an issue where the type inherits JsonTypeInfo from some
ancestor but doesn't have its own JsonTypeIdResolver.  note that this
did break a (seemingly minor?) bit of functionality in the
EncoderDecoderTestGwt class, hopefully that is ok.